requirements.txt
appdirs==1.4.3
appnope==0.1.0
astroid==2.4.0
attrs==19.3.0
backcall==0.1.0
black==19.10b0
bleach==3.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.1.0
jupyterlab-server==1.1.0
kaggle==1.5.6
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
olefile==0.46
opencv-python==4.2.0.34
pandocfilters==1.4.2
parso==0.6.2
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pylint==2.5.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-slugify==4.0.0
pyzmq==18.1.1
qtconsole==4.7.2
QtPy==1.9.0
regex==2020.4.4
requests==2.23.0
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
toml==0.10.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.24.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==2.2.0
